+++ calyx/opt/src/default_passes.rs
@@ -4,8 +4,8 @@ use crate::passes::{
     AddGuard, Canonicalize, CellShare, ClkInsertion, CollapseControl, CombProp,
     CompileInvoke, CompileRepeat, CompileStatic, ComponentInliner,
     ConstantPortProp, DataPathInfer, DeadAssignmentRemoval, DeadCellRemoval,
-    DeadGroupRemoval, DefaultAssigns, Externalize, FSMAnnotator, GoInsertion,
-    GroupToInvoke, GroupToSeq, InferShare, LowerGuards, MergeAssign, Papercut,
+    DeadGroupRemoval, DefaultAssigns, Externalize, GoInsertion, GroupToInvoke,
+    GroupToSeq, InferShare, LowerGuards, MergeAssign, Papercut,
     ProfilerInstrumentation, RemoveIds, ResetInsertion, SimplifyStaticGuards,
     SimplifyWithControl, StaticFSMAllocation, StaticFSMOpts, StaticInference,
     StaticInliner, StaticPromotion, StaticRepeatFSMAllocation,
