+++ calyx/opt/src/passes/mod.rs
@@ -13,7 +13,6 @@ mod dead_assignment_removal;
 mod dead_cell_removal;
 mod dead_group_removal;
 mod externalize;
-mod fsm_annotator;
 mod go_insertion;
 mod group_to_invoke;
 mod group_to_seq;
@@ -85,7 +84,6 @@ pub use unroll_bound::UnrollBounded;
 // pub use simplify_guards::SimplifyGuards;
 pub use add_guard::AddGuard;
 pub use default_assigns::DefaultAssigns;
-pub use fsm_annotator::FSMAnnotator;
 pub use synthesis_papercut::SynthesisPapercut;
 pub use top_down_compile_control::TopDownCompileControl;
 pub use well_formed::WellFormed;
