+++ calyx/opt/src/passes_experimental/fsm_annotator.rs
@@ -1,489 +1,296 @@
-use crate::traversal::{
-    Action, ConstructVisitor, Named, ParseVal, PassOpt, VisResult, Visitor,
+use crate::{
+    analysis::{
+        AcyclicAnnotation, AnnotatedControlNode, RepeatNodeAnnotation,
+        WhileNodeAnnotation,
+    },
+    traversal::{Action, ConstructVisitor, Named, VisResult, Visitor},
 };
 use calyx_ir::{self as ir};
+use calyx_utils::CalyxResult;
 
-/// From the perspective of a parent control node, the annotation `@NUM_STATES(n)`
-/// on one of its child nodes means that the parent should allocate `n` states
-/// in order to implement the child's schedule.
-const NUM_STATES: ir::Attribute =
-    ir::Attribute::Internal(ir::InternalAttr::NUM_STATES);
-
-/// From the perspective of a parent control node, if the attribute `@ACYCLIC` is
-/// present on one of its children, then the following implication is guaranteed:
-///
-/// If the child's FSM has an @UNROLL or @INLINE attribute, then
-/// the states implementing the child's schedule should have the property of
-/// having one state for every cycle.
-///
-/// If the attribute `@ACYCLIC` is not present on a child node, then this simply
-/// means that the states implementing the child's control has a backedge.
-///
-/// For now, it is undefined behavior for one child to have both `@ACYCLIC` and
-/// `@OFFLOAD` attributes attached.
-const ACYCLIC: ir::Attribute =
-    ir::Attribute::Internal(ir::InternalAttr::ACYCLIC);
-
-fn is_acyclic(ctrl: &ir::StaticControl) -> bool {
-    ctrl.get_attribute(ACYCLIC).is_some()
-}
-
-fn get_num_states(ctrl: &ir::Control) -> u64 {
-    ctrl.get_attribute(NUM_STATES).unwrap()
-}
-
-fn get_num_states_static(ctrl: &ir::StaticControl) -> u64 {
-    ctrl.get_attribute(NUM_STATES).unwrap()
-}
-
-/// Given the attributes field of a control node, insert the annotations derived
-/// in this pass into the `control {...}` section of a Calyx program.
-fn transfer_attributes(
-    attributes_map: &mut ir::Attributes,
-    FSMImplementation {
-        num_states,
-        acyclic,
-        attr,
-    }: &FSMImplementation,
-) {
-    attributes_map.insert(NUM_STATES, *num_states);
-    let ir_attribute = match attr {
-        ControlNodeAnnotation::Inline(inline_impl) => match inline_impl {
-            InlineImplementation::StandardInline => ir::InternalAttr::INLINE,
-            InlineImplementation::Unroll => ir::InternalAttr::UNROLL,
-        },
-        ControlNodeAnnotation::Offload => ir::InternalAttr::OFFLOAD,
-    };
-    attributes_map.insert(ir::Attribute::Internal(ir_attribute), 1);
-    if *acyclic {
-        attributes_map.insert(ACYCLIC, 1);
-    }
-}
-
-/// Encodes the possible FSM implementations of a control node. From the perspective
-/// of a parent FSM, the states implementing each of its children can
-/// either be brought into the parent FSM (i.e. inlined) or can be outsourced
-/// to another FSM (offloaded).
-enum ControlNodeAnnotation {
-    Inline(InlineImplementation),
-    Offload,
-}
-
-/// Encodes the possible ways for a control node's FSM to be inlined into that of its parent.
-/// Here, `Unroll` can only be used on ir::Repeat or ir::StaticRepeat nodes. In the event
-/// that `StandardInline` is used on one of these nodes, then, within the parent
-/// FSM, there will be a backedge from the bottom to the top of the repeat.
-///
-/// The annotation `StandardInline` is used to indicate inlining for every other
-/// control node.
-enum InlineImplementation {
-    Unroll,
-    StandardInline,
-}
-
-/// An instance of this struct is computed for every control node. It will be used
-/// by its parent control node.
-struct FSMImplementation {
-    num_states: u64,
-    acyclic: bool,
-    attr: ControlNodeAnnotation,
-}
-
-trait FSMPolicy {
-    /// Given a control node, returns a number of states allocated for the FSM,
-    /// whether backedges exist in the FSM, and an attribute for this node.
-    fn policy(ctrl: &mut Self, child_fsm_cutoff: u64) -> FSMImplementation;
-}
-
-impl FSMPolicy for ir::FSMEnable {
-    fn policy(_: &mut Self, _: u64) -> FSMImplementation {
-        FSMImplementation {
-            num_states: 1,
-            acyclic: false,
-            attr: ControlNodeAnnotation::Offload,
-        }
-    }
-}
-
-impl FSMPolicy for ir::StaticEnable {
-    fn policy(
-        ctrl: &mut ir::StaticEnable,
-        child_fsm_cutoff: u64,
-    ) -> FSMImplementation {
-        let (num_states, acyclic) = {
-            let latency = ctrl.group.borrow().get_latency();
-            if latency < child_fsm_cutoff {
-                (latency, true)
-            } else {
-                (1, false)
-            }
-        };
-
-        FSMImplementation {
-            num_states,
-            acyclic,
-            attr: ControlNodeAnnotation::Inline(
-                InlineImplementation::StandardInline,
-            ),
-        }
-    }
-}
-
-impl FSMPolicy for ir::StaticSeq {
-    fn policy(ctrl: &mut ir::StaticSeq, _: u64) -> FSMImplementation {
-        let (num_states, acyclic) =
-            ctrl.stmts
-                .iter()
-                .fold((0, true), |(num_states, acyclic), stmt| {
-                    let stmt_latency = get_num_states_static(stmt);
-                    let stmt_acyclic = is_acyclic(stmt);
-                    (num_states + stmt_latency, acyclic && stmt_acyclic)
-                });
-
-        FSMImplementation {
-            num_states,
-            acyclic,
-            attr: ControlNodeAnnotation::Inline(
-                InlineImplementation::StandardInline,
-            ),
-        }
-    }
-}
-
-impl FSMPolicy for ir::StaticPar {
-    fn policy(ctrl: &mut ir::StaticPar, _: u64) -> FSMImplementation {
-        let (num_states, acyclic, attr) = if ctrl.stmts.iter().all(is_acyclic) {
-            (
-                ctrl.latency,
-                true,
-                ControlNodeAnnotation::Inline(
-                    InlineImplementation::StandardInline,
-                ),
-            )
-        } else {
-            (1, false, ControlNodeAnnotation::Offload)
-        };
-
-        FSMImplementation {
-            num_states,
-            acyclic,
-            attr,
-        }
-    }
-}
-
-impl FSMPolicy for ir::StaticRepeat {
-    fn policy(
-        ctrl: &mut ir::StaticRepeat,
-        child_fsm_cutoff: u64,
-    ) -> FSMImplementation {
-        let (body_num_states, body_is_acyclic) =
-            (get_num_states_static(&ctrl.body), is_acyclic(&ctrl.body));
-        let unrolled_num_states = ctrl.num_repeats * body_num_states;
-        let (num_states, acyclic, attr) =
-            if body_is_acyclic && (unrolled_num_states < child_fsm_cutoff) {
-                (
-                    unrolled_num_states,
-                    true,
-                    ControlNodeAnnotation::Inline(InlineImplementation::Unroll),
-                )
-            } else if body_num_states < child_fsm_cutoff {
-                (
-                    body_num_states,
-                    false,
-                    ControlNodeAnnotation::Inline(
-                        InlineImplementation::StandardInline,
-                    ),
-                )
-            } else {
-                (1, false, ControlNodeAnnotation::Offload)
-            };
-
-        FSMImplementation {
-            num_states,
-            acyclic,
-            attr,
-        }
-    }
-}
-
-impl FSMPolicy for ir::StaticIf {
-    fn policy(ctrl: &mut ir::StaticIf, _: u64) -> FSMImplementation {
-        let (num_states, acyclic, attr) =
-            if is_acyclic(&ctrl.tbranch) && is_acyclic(&ctrl.fbranch) {
-                (
-                    ctrl.latency,
-                    true,
-                    ControlNodeAnnotation::Inline(
-                        InlineImplementation::StandardInline,
-                    ),
-                )
-            } else {
-                (1, false, ControlNodeAnnotation::Offload)
-            };
-
-        FSMImplementation {
-            num_states,
-            acyclic,
-            attr,
-        }
-    }
-}
-
-impl FSMPolicy for ir::Enable {
-    fn policy(_ctrl: &mut ir::Enable, _: u64) -> FSMImplementation {
-        FSMImplementation {
-            num_states: 1,
-            acyclic: false,
-            attr: ControlNodeAnnotation::Inline(
-                InlineImplementation::StandardInline,
-            ),
-        }
-    }
-}
-
-impl FSMPolicy for ir::Seq {
-    fn policy(ctrl: &mut ir::Seq, _: u64) -> FSMImplementation {
-        FSMImplementation {
-            num_states: ctrl.stmts.iter().map(get_num_states).sum(),
-            acyclic: false,
-            attr: ControlNodeAnnotation::Inline(
-                InlineImplementation::StandardInline,
-            ),
-        }
-    }
-}
-
-impl FSMPolicy for ir::Par {
-    fn policy(_ctrl: &mut ir::Par, _: u64) -> FSMImplementation {
-        FSMImplementation {
-            num_states: 1,
-            acyclic: false,
-            attr: ControlNodeAnnotation::Offload,
-        }
-    }
-}
-
-impl FSMPolicy for ir::If {
-    fn policy(_ctrl: &mut ir::If, _: u64) -> FSMImplementation {
-        FSMImplementation {
-            num_states: 1,
-            acyclic: false,
-            attr: ControlNodeAnnotation::Offload,
-        }
-    }
-}
-
-impl FSMPolicy for ir::While {
-    fn policy(ctrl: &mut ir::While, _: u64) -> FSMImplementation {
-        FSMImplementation {
-            num_states: ctrl.body.get_attribute(NUM_STATES).unwrap(),
-            acyclic: false,
-            attr: ControlNodeAnnotation::Inline(
-                InlineImplementation::StandardInline,
-            ),
-        }
-    }
-}
-
-impl FSMPolicy for ir::Repeat {
-    fn policy(ctrl: &mut ir::Repeat, _: u64) -> FSMImplementation {
-        FSMImplementation {
-            num_states: ctrl.body.get_attribute(NUM_STATES).unwrap(),
-            acyclic: false,
-            attr: ControlNodeAnnotation::Inline(
-                InlineImplementation::StandardInline,
-            ),
-        }
-    }
-}
-
-pub struct FSMAnnotator {
-    child_fsm_cutoff: u64,
-}
+const NODE_ID: ir::Attribute =
+    ir::Attribute::Internal(ir::InternalAttr::NODE_ID);
+pub struct FSMAnnotator {}
 
 impl Named for FSMAnnotator {
     fn name() -> &'static str {
         "fsm-annotator"
     }
-
     fn description() -> &'static str {
         "annotate a control program, determining how FSMs should be allocated"
     }
-
-    fn opts() -> Vec<PassOpt> {
-        vec![PassOpt::new(
-            "child-fsm-cutoff",
-            "The maximum number of states a child FSM can have, before it is offloaded",
-            ParseVal::Num(300),
-            PassOpt::parse_num,
-        )]
-    }
 }
 impl ConstructVisitor for FSMAnnotator {
-    fn from(ctx: &ir::Context) -> calyx_utils::CalyxResult<Self> {
-        let opts = Self::get_opts(ctx);
-        Ok(FSMAnnotator {
-            child_fsm_cutoff: opts[&"child-fsm-cutoff"]
-                .pos_num()
-                .expect("requires non-negative OHE cutoff parameter"),
-        })
+    fn from(_: &ir::Context) -> CalyxResult<Self> {
+        Ok(FSMAnnotator {})
     }
     fn clear_data(&mut self) {}
 }
 
+impl<'a> FSMAnnotator {
+    /// Code abstraction: given an iterator over mutable references to
+    /// ir::Control nodes, apply `Self::update_node_with_id`.
+    fn map_update_across<I>(
+        mut iter: I,
+        id: u64,
+        (attr, attr_val): (ir::Attribute, u64),
+    ) -> Option<()>
+    where
+        I: Iterator<Item = &'a mut ir::Control>,
+    {
+        if iter.any(|stmt| {
+            matches!(
+                Self::update_node_with_id(stmt, id, (attr, attr_val),),
+                Some(())
+            )
+        }) {
+            Some(())
+        } else {
+            None
+        }
+    }
+
+    /// Code abstraction: given an iterator over mutable references to
+    /// ir::StaticControl nodes, apply `Self::update_static_node_with_id`.
+    fn map_update_across_static<I>(
+        mut iter: I,
+        id: u64,
+        (attr, attr_val): (ir::Attribute, u64),
+    ) -> Option<()>
+    where
+        I: Iterator<Item = &'a mut ir::StaticControl>,
+    {
+        if iter.any(|stmt| {
+            matches!(
+                Self::update_static_node_with_id(stmt, id, (attr, attr_val)),
+                Some(())
+            )
+        }) {
+            Some(())
+        } else {
+            None
+        }
+    }
+}
+
+impl FSMAnnotator {
+    /// Given a mutable reference to an `ir::Control` tree, update a specifc node
+    /// (identified by `id`) with an attribute.
+    fn update_node_with_id(
+        ctrl: &mut ir::Control,
+        id: u64,
+        (attr, attr_val): (ir::Attribute, u64),
+    ) -> Option<()> {
+        match ctrl.get_attribute(NODE_ID) {
+            Some(node_id_val) if node_id_val == id => {
+                ctrl.insert_attribute(attr, attr_val);
+                Some(())
+            }
+            _ => match ctrl {
+                ir::Control::Empty(_)
+                | ir::Control::Enable(_)
+                | ir::Control::FSMEnable(_) => None,
+                ir::Control::While(whle) => Self::update_node_with_id(
+                    &mut whle.body,
+                    id,
+                    (attr, attr_val),
+                ),
+                ir::Control::Repeat(repeat) => Self::update_node_with_id(
+                    &mut repeat.body,
+                    id,
+                    (attr, attr_val),
+                ),
+                ir::Control::Seq(seq) => Self::map_update_across(
+                    seq.stmts.iter_mut(),
+                    id,
+                    (attr, attr_val),
+                ),
+                ir::Control::Par(par) => Self::map_update_across(
+                    par.stmts.iter_mut(),
+                    id,
+                    (attr, attr_val),
+                ),
+                ir::Control::If(dif) => Self::map_update_across(
+                    vec![dif.tbranch.as_mut(), dif.fbranch.as_mut()]
+                        .into_iter(),
+                    id,
+                    (attr, attr_val),
+                ),
+                ir::Control::Static(sctrl) => Self::update_static_node_with_id(
+                    sctrl,
+                    id,
+                    (attr, attr_val),
+                ),
+                ir::Control::Invoke(_) => {
+                    unreachable!("Invoke nodes should have been compiled away")
+                }
+            },
+        }
+    }
+
+    /// Given a mutable reference to an `ir::StaticControl` tree, update a specifc
+    /// node (identified by `id`) with an attribute.
+    fn update_static_node_with_id(
+        sctrl: &mut ir::StaticControl,
+        id: u64,
+        (attr, attr_val): (ir::Attribute, u64),
+    ) -> Option<()> {
+        match sctrl.get_attribute(NODE_ID) {
+            Some(node_id_val) if node_id_val == id => {
+                sctrl.insert_attribute(attr, attr_val);
+                Some(())
+            }
+            _ => match sctrl {
+                ir::StaticControl::Empty(_) | ir::StaticControl::Enable(_) => {
+                    None
+                }
+                ir::StaticControl::Seq(sseq) => Self::map_update_across_static(
+                    sseq.stmts.iter_mut(),
+                    id,
+                    (attr, attr_val),
+                ),
+                ir::StaticControl::Par(spar) => Self::map_update_across_static(
+                    spar.stmts.iter_mut(),
+                    id,
+                    (attr, attr_val),
+                ),
+                ir::StaticControl::If(sif) => Self::map_update_across_static(
+                    vec![sif.tbranch.as_mut(), sif.fbranch.as_mut()]
+                        .into_iter(),
+                    id,
+                    (attr, attr_val),
+                ),
+                ir::StaticControl::Repeat(srep) => {
+                    Self::update_static_node_with_id(
+                        &mut srep.body,
+                        id,
+                        (attr, attr_val),
+                    )
+                }
+                ir::StaticControl::Invoke(_) => {
+                    unreachable!("Invoke nodes should have been compiled away")
+                }
+            },
+        }
+    }
+
+    /// Given a reference to the data structure on which analysis occurs,
+    /// translate these annotations into the Calyx IR data structure.
+    fn project_onto_control(
+        abstract_control: &AnnotatedControlNode,
+        control: &mut ir::Control,
+    ) {
+        match abstract_control {
+            AnnotatedControlNode::StaticSeq { stmts, .. }
+            | AnnotatedControlNode::DynamicSeq { stmts, .. } => stmts
+                .iter()
+                .for_each(|stmt| Self::project_onto_control(stmt, control)),
+            AnnotatedControlNode::StaticPar {
+                id,
+                threads,
+                acyclic,
+                ..
+            } => {
+                if matches!(acyclic.unwrap(), AcyclicAnnotation::True) {
+                    let attr =
+                        ir::Attribute::Internal(ir::InternalAttr::LOCKSTEP);
+                    Self::update_node_with_id(control, *id, (attr, 1));
+                }
+                threads.iter().for_each(|thread| {
+                    Self::project_onto_control(thread, control)
+                })
+            }
+            AnnotatedControlNode::StaticIf {
+                id,
+                true_thread,
+                false_thread,
+                acyclic,
+                ..
+            } => {
+                if matches!(acyclic.unwrap(), AcyclicAnnotation::True) {
+                    let attr =
+                        ir::Attribute::Internal(ir::InternalAttr::LOCKSTEP);
+                    Self::update_node_with_id(control, *id, (attr, 1));
+                }
+                Self::project_onto_control(true_thread, control);
+                Self::project_onto_control(false_thread, control);
+            }
+            AnnotatedControlNode::StaticRepeat {
+                id,
+                body,
+                annotation,
+                ..
+            }
+            | AnnotatedControlNode::DynamicRepeat {
+                id,
+                body,
+                annotation,
+                ..
+            } => {
+                let body_attr =
+                    ir::Attribute::Internal(match annotation.unwrap() {
+                        RepeatNodeAnnotation::Inline => {
+                            ir::InternalAttr::INLINE
+                        }
+                        RepeatNodeAnnotation::Offload => {
+                            ir::InternalAttr::OFFLOAD
+                        }
+                        RepeatNodeAnnotation::Unroll => {
+                            ir::InternalAttr::UNROLL
+                        }
+                    });
+                Self::update_node_with_id(control, *id, (body_attr, 1));
+                Self::project_onto_control(body, control);
+            }
+            AnnotatedControlNode::DynamicWhile {
+                id,
+                body,
+                annotation,
+                ..
+            } => {
+                let body_attr =
+                    ir::Attribute::Internal(match annotation.unwrap() {
+                        WhileNodeAnnotation::Inline => ir::InternalAttr::INLINE,
+                        WhileNodeAnnotation::Offload => {
+                            ir::InternalAttr::OFFLOAD
+                        }
+                    });
+                Self::update_node_with_id(control, *id, (body_attr, 1));
+                Self::project_onto_control(body, control);
+            }
+            AnnotatedControlNode::DynamicPar { threads, .. } => threads
+                .iter()
+                .for_each(|thread| Self::project_onto_control(thread, control)),
+            AnnotatedControlNode::DynamicIf {
+                true_thread,
+                false_thread,
+                ..
+            } => {
+                Self::project_onto_control(true_thread, control);
+                Self::project_onto_control(false_thread, control);
+            }
+
+            _ => (),
+        }
+    }
+}
+
 impl Visitor for FSMAnnotator {
-    fn empty(
+    fn start(
         &mut self,
-        s: &mut ir::Empty,
-        _comp: &mut ir::Component,
+        comp: &mut ir::Component,
         _sigs: &ir::LibrarySignatures,
         _comps: &[ir::Component],
     ) -> VisResult {
-        s.attributes.insert(NUM_STATES, 0);
-        Ok(Action::Continue)
-    }
-    fn enable(
-        &mut self,
-        s: &mut ir::Enable,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::Enable::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
+        // Build the abstract data structure and analyze nodes at which
+        // the annotations should take place
+        let mut node: AnnotatedControlNode =
+            AnnotatedControlNode::from(&mut *comp.control.borrow_mut());
+        node.post_order_analysis();
 
-    fn static_enable(
-        &mut self,
-        s: &mut ir::StaticEnable,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::StaticEnable::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
+        // translate these annotations onto ir::Control
+        Self::project_onto_control(&node, &mut comp.control.borrow_mut());
 
-    fn fsm_enable(
-        &mut self,
-        s: &mut ir::FSMEnable,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::FSMEnable::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
-
-    fn finish_static_seq(
-        &mut self,
-        s: &mut ir::StaticSeq,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::StaticSeq::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
-
-    fn finish_static_par(
-        &mut self,
-        s: &mut ir::StaticPar,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::StaticPar::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
-
-    fn finish_static_if(
-        &mut self,
-        s: &mut ir::StaticIf,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::StaticIf::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
-
-    fn finish_static_repeat(
-        &mut self,
-        s: &mut ir::StaticRepeat,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::StaticRepeat::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
-
-    fn finish_seq(
-        &mut self,
-        s: &mut ir::Seq,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::Seq::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
-
-    fn finish_par(
-        &mut self,
-        s: &mut ir::Par,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::Par::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
-
-    fn finish_if(
-        &mut self,
-        s: &mut ir::If,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::If::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
-
-    fn finish_while(
-        &mut self,
-        s: &mut ir::While,
-        _comp: &mut ir::Component,
-        _sigs: &ir::LibrarySignatures,
-        _comps: &[ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::While::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
-        Ok(Action::Continue)
-    }
-
-    fn finish_repeat(
-        &mut self,
-        s: &mut calyx_ir::Repeat,
-        _comp: &mut calyx_ir::Component,
-        _sigs: &calyx_ir::LibrarySignatures,
-        _comps: &[calyx_ir::Component],
-    ) -> VisResult {
-        let fsm_impl = ir::Repeat::policy(s, self.child_fsm_cutoff);
-        transfer_attributes(&mut s.attributes, &fsm_impl);
         Ok(Action::Continue)
     }
 }
